SA-373: Tidy the attribute authority stuff in KAS
