+++ containers/kas/kas_app/scripts/entrypoint
@@ -19,7 +19,7 @@ else
 fi
 
 export CA_CERT_PATH
-export EAS_CERTIFICATE
+export ATTR_AUTHORITY_CERTIFICATE
 export KAS_EC_SECP256R1_PRIVATE_KEY
 export KAS_EC_SECP256R1_CERTIFICATE
 export KAS_PRIVATE_KEY
