+++ containers/kas/kas_app/scripts/test.env
@@ -1,6 +1,6 @@
 CERTS_DIR="../../../certs"
 . "${CERTS_DIR}/.env"
-export EAS_CERTIFICATE
+export ATTR_AUTHORITY_CERTIFICATE
 export KAS_CERTIFICATE
 export KAS_EC_SECP256R1_CERTIFICATE
 export KAS_EC_SECP256R1_PRIVATE_KEY
