+++ containers/kas/kas_app/tdf3_kas_app/plugins/opentdf_attr_authority_plugin_test.py
@@ -1,4 +1,4 @@
-"""Test the EAS connector."""
+"""Test the OpenTDF attribute authority plugin."""
 
 import pytest
 
@@ -8,7 +8,7 @@ import requests
 
 from tdf3_kas_core.errors import InvalidAttributeError
 from tdf3_kas_core.errors import RequestTimeoutError
-from .eas_connector import EASConnector
+from .opentdf_attr_authority_plugin import OpenTDFAttrAuthorityPlugin
 
 HOST = "http://localhost:4010"
 NAMESPACES = [
@@ -24,20 +24,16 @@ ATTRIBUTES = {
 }
 
 
-def test_eas_connector_constructor():
-    connector = EASConnector(HOST)
-    assert isinstance(connector, EASConnector)
-    assert connector._host == HOST
-    assert connector._headers == {
-        "Content-Type": "application/json",
-    }
-    assert connector._requests_timeout == 10
+def test_otdf_plugin_constructor():
+    actual = OpenTDFAttrAuthorityPlugin(HOST)
+    assert isinstance(actual, OpenTDFAttrAuthorityPlugin)  # my kingdom for static typing.
+    assert actual._host == HOST
 
 
 @patch.object(requests, "post", return_value=Mock(status_code=404))
 def test_fetch_attributes_404(mock_request):
-    connector = EASConnector(HOST)
-    result = connector.fetch_attributes([])
+    actual = OpenTDFAttrAuthorityPlugin(HOST)
+    result = actual.fetch_attributes([])
     assert result is None
 
 
@@ -55,27 +51,27 @@ def mocked_requests_response(*args, **kwargs):
 
 @patch.object(requests, "post", side_effect=mocked_requests_response)
 def test_fetch_attributes_200_stautus(mock_request):
-    connector = EASConnector(HOST)
-    attributes = connector.fetch_attributes(NAMESPACES)
+    actual = OpenTDFAttrAuthorityPlugin(HOST)
+    attributes = actual.fetch_attributes(NAMESPACES)
     assert attributes == ATTRIBUTES
 
 
 @patch.object(requests, "post", side_effect=requests.exceptions.ReadTimeout)
 def test_fetch_attributes_read_timeout_exception(mock_request):
-    connector = EASConnector(HOST)
+    actual = OpenTDFAttrAuthorityPlugin(HOST)
     with pytest.raises(RequestTimeoutError):
-        connector.fetch_attributes(NAMESPACES)
+        actual.fetch_attributes(NAMESPACES)
 
 
 @patch.object(requests, "post", side_effect=requests.exceptions.ConnectTimeout)
 def test_fetch_attributes_connect_timeout_exception(mock_request):
-    connector = EASConnector(HOST)
+    actual = OpenTDFAttrAuthorityPlugin(HOST)
     with pytest.raises(RequestTimeoutError):
-        connector.fetch_attributes(NAMESPACES)
+        actual.fetch_attributes(NAMESPACES)
 
 
 @patch.object(requests, "post", side_effect=requests.exceptions.RequestException)
 def test_fetch_attributes_request_exception(mock_request):
-    connector = EASConnector(HOST)
+    actual = OpenTDFAttrAuthorityPlugin(HOST)
     with pytest.raises(InvalidAttributeError):
-        connector.fetch_attributes(NAMESPACES)
+        actual.fetch_attributes(NAMESPACES)
