+++ deployments/pki_local/nginx.ingress.yaml
@@ -26,13 +26,6 @@ spec:
     - host: opentdf.local
       http:
         paths:
-          - path: /eas/(.*)
-            pathType: ImplementationSpecific
-            backend:
-              service:
-                name: eas
-                port:
-                  name: http
           - path: /kas/(.*)
             pathType: ImplementationSpecific
             backend:
