+++ tests/performance-test/docker-compose.yml
@@ -102,8 +102,8 @@ services:
       - statsd
       - aa
     environment:
-      - EAS_HOST=http://aa:4020
-      - EAS_CERTIFICATE
+      - ATTR_AUTHORITY_HOST=http://aa:4020
+      - ATTR_AUTHORITY_CERTIFICATE
       - FLASK_DEBUG=True
       - FLASK_TESTING=True
       - STATSD_HOST=statsd
